feed dic
